% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextTfFeatures.R
\name{contextTfFeatures}
\alias{contextTfFeatures}
\title{Cellular context and transcription factor-specific features}
\usage{
contextTfFeatures(
  mae,
  tfName,
  addLabels = FALSE,
  whichCol = c("All", "OnlyTrain", "Col"),
  colSel = NULL,
  features = c("Inserts", "Weighted_Inserts", "Cofactor_Inserts", "ChromVAR_Scores",
    "Cofactor_ChromVAR_Scores"),
  annoCol = "context",
  insertionProfile = NULL,
  subSample = 10000,
  aggregationFun = sum,
  seed = 42,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{mae}{\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} as construced by \code{\link[=prepData]{prepData()}} containing Motif, ATAC-, ChIP-seq,
site-specific features as obtained by \code{\link[=siteFeatures]{siteFeatures()}} and transcription factor-specific features as obtained by \code{\link[=tfFeatures]{tfFeatures()}}.}

\item{tfName}{Name of transcription factor to compute features for.}

\item{addLabels}{Should ChIP-seq peak labels be added to the features}

\item{whichCol}{Should features be calculated for all cellular contexts (\code{"All"}), only the training data (\code{"OnlyTrain"})
or only for some specific cellular contexts (\code{"Col"}) specified in \code{colSel}.}

\item{colSel}{If \code{whichCol="colSel"}, name of the cellular context to compute the features for.}

\item{features}{Names of features to be added. Can be all or some of "Inserts", "Weighted_Inserts", "Cofactor_Inserts".
See \link{listFeatures} for an overview of the features.}

\item{annoCol}{Name of column indicating cellular contexts in colData.}

\item{insertionProfile}{Pre-computed insertion footprint profile for the specified transcription factor.
Needs to contain coordinate (chr/seqnames, start, end) columns and weight column (termed "w").}

\item{subSample}{If "ChromVAR_Scores" amongst features, the number of sites to subsample for computing the scores. If \code{NULL} no subsampling performed.}

\item{aggregationFun}{function (e.g. mean, median, sum) used to aggregate features across the rowRanges of experiments of the
provided \link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} object.}

\item{seed}{Integer value for setting the seed for random number generation with \link[base:Random]{base::set.seed}.}

\item{BPPARAM}{Parallel back-end to be used. Passed to \code{\link[BiocParallel:bpmapply]{BiocParallel::bpmapply()}} & \code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}}.}

\item{...}{Arguments passed to \link{getInsertionProfiles} and \link[chromVAR:getBackgroundPeaks]{chromVAR::getBackgroundPeaks}.}
}
\value{
\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} object with an experiment containing transcription factor- and cellular context-specific features added to \link[MultiAssayExperiment:MultiAssayExperiment-methods]{MultiAssayExperiment::experiments}.
If already an experiment of this feature group exists, columns are added to it.
}
\description{
Adds an experiment with features specific for the specified transcription factor and the cellular-contexts it is covered for,
such as Tn5 insertion around motif matches and footprint profiles.
}
