% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordFeatures.R
\name{coordFeatures}
\alias{coordFeatures}
\title{Adds an experiment with features specific for each site, such as CpG-content, conservation scores etc.}
\usage{
coordFeatures(
  mae,
  aggFun = max,
  annoData = NULL,
  scoreCols = NULL,
  features = c("Sequence", "Width", "Annot"),
  ...
)
}
\arguments{
\item{mae}{MultiAssayExperiment as construced by \code{\link[=prepData]{prepData()}} containing ATAC- and ChIP-seq data.}

\item{aggFun}{Aggregation function for aggregating genomic scores overlapping the rowRanges of the provided MultiAssayExperiment object.}

\item{annoData}{Further data to be aggregated across the rowRanges of the provided MultiAssayExperiment object.
List of paths pointing to .bed / .bam files or data.frames / data.tables or GRanges containing chr, start and end columns and score columns.}

\item{scoreCols}{Column names of the scores of annoData to be aggregated.}

\item{features}{Names of features to be added. Can be all or some of "Sequence", "Width", "Annot".}
}
\description{
Adds an experiment with features specific for each site, such as CpG-content, conservation scores etc.
}
