% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\docType{data}
\name{example_coords}
\alias{example_coords}
\title{Example Data: Coordinates}
\format{
\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object
}
\description{
Example genomic coordinates to construct a custom MultiAssayExperiment object and to compute features for TF-Binding predictions for.
The sample data contains 100 genomic ranges stored in GRanges object.
}
\keyword{GRanges}
