% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\docType{data}
\name{example_pwms}
\alias{example_pwms}
\title{Example Data: Motifs}
\format{
\link[TFBSTools:XMatrixList-class]{TFBSTools::PWMatrixList} object
}
\description{
Example motifs of JUN and CTCF as PWMatrixList. Can be used as an example input for prepMotifs().
}
\keyword{PWMatrixList}
