% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{getContexts}
\alias{getContexts}
\title{Get cellular contexts}
\usage{
getContexts(mae, tfName = NULL, which = c("ChIP", "ATAC", "Both"))
}
\arguments{
\item{mae}{\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} as construced by \code{\link[=prepData]{prepData()}} containing Motif, ATAC- and ChIP-seq data.}

\item{tfName}{Name of transcription factor to get cellular contexts for.
If \code{tfName=NULL} all celullar contexts of the requested modalities (\code{which}), will be provided.}

\item{which}{Which modality to get the cellular contexts names from.
If \code{which=both} only cellular contexts which have both ATAC and ChIP-seq data (for the specified TF, in case \code{tfName} is not NULL) will be returned.}
}
\description{
Provides the names of cellular contexts for data of interest, e.g. a TF or modality.
}
