% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureMatrix.R
\name{getFeatureMatrix}
\alias{getFeatureMatrix}
\title{Feature matrix construction}
\usage{
getFeatureMatrix(
  mae,
  tfName,
  addLabels = TRUE,
  whichCol = c("All", "OnlyTrain", "Col"),
  colNorm = TRUE,
  convertInteger = FALSE,
  saveHdf5 = TRUE,
  outDir = NULL,
  prefix = NULL,
  annoCol = "context",
  BPPARAM = SerialParam()
)
}
\arguments{
\item{mae}{\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} as construced by \code{\link[=prepData]{prepData()}} containing Motif, ATAC-, ChIP-seq,
site-specific features as obtained by \code{\link[=siteFeatures]{siteFeatures()}}, transcription factor-specific features as obtained by \code{\link[=tfFeatures]{tfFeatures()}}
and transcription-factor and cellular-context specific features as obtained by \code{\link[=contextTfFeatures]{contextTfFeatures()}}.}

\item{tfName}{Name of transcription factor to compute features for.}

\item{addLabels}{Should ChIP-seq peak labels be added to the feature matrix.}

\item{whichCol}{Should feature matrix be calculated for all cellular contexts (\code{"All"}), only the training data (\code{"OnlyTrain"})
or only for some specific cellular contexts (\code{"Col"}) specified in \code{colSel}.}

\item{colNorm}{If cellular-context specific features should be column-normalized.}

\item{convertInteger}{If feature matrix should be converted to integer (to lower memory footprint).}

\item{saveHdf5}{If feature matrix should be saved as HDF5 file.}

\item{outDir}{Directory to save HDF5 file to.}

\item{prefix}{Prefix added to filename of feature matrix in case saved as HDF5 file.}

\item{annoCol}{Name of column indicating cellular contexts in colData.}

\item{BPPARAM}{Parallel back-end to be used. Passed to \code{\link[BiocParallel:bpmapply]{BiocParallel::bpmapply()}}.}
}
\value{
Feature Matrix either as \link[Matrix:Matrix-class]{Matrix::Matrix} or as \link[HDF5Array:HDF5Array-class]{HDF5Array::HDF5Array} (if \code{saveHdf5=TRUE}).
}
\description{
Compiles feature matrix based on pre-computed features stored in experiments of the provided \link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} object.
}
