% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInsertionProfiles.R
\name{getInsertionProfiles}
\alias{getInsertionProfiles}
\title{Tn5 insertion counting}
\usage{
getInsertionProfiles(
  atacData,
  motifRanges,
  margin = 200,
  shift = FALSE,
  calcProfile = TRUE,
  profiles = NULL,
  symmetric = FALSE,
  stranded = FALSE
)
}
\arguments{
\item{atacData}{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}, \link[data.table:data.table]{data.table::data.table}, data.frames or paths to .bed /. bam files
containing ATAC-seq fragment coordinates (i.e. chr/seqnames, start, end and optionally a strand column).
If it contains a column named "sample" insertion counts and profiles will be computed for each sample.}

\item{motifRanges}{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object containing coordinates of motif-matches.}

\item{margin}{Margin around motif-matches to consider for computing Tn5 insertion events}

\item{shift}{If Tn5 insertion bias should be considered (only if strand column is provided).}

\item{calcProfile}{If Insertion footprint profiles should be computed.}

\item{profiles}{Pre-computed insertion footprint profile to compute weighted insertion counts in case \code{calcProfile=FALSE}.
Needs to contain coordinate (chr/seqnames, start, end) columns and weight column (termed "w").}

\item{symmetric}{If transcription factor footprint profiles should be symmetric around the motif matches. Only used if \code{calcProfile=TRUE}.}
}
\value{
\link[data.table:data.table]{data.table::data.table} containing insertion counts within and in margins around motif matches and weighted insertion counts in case
an insertion profile is provided or if \code{calcProfile=TRUE}.
}
\description{
Counts Tn5 insertions around provided motif-matches.
If requested also computes insertion footprint profiles and weighted insertion counts.
}
