% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoadH5mae.R
\name{getMaeH5paths}
\alias{getMaeH5paths}
\title{getMaeH5paths - extract HDF5 file paths from MAE}
\usage{
getMaeH5paths(x)
}
\arguments{
\item{x}{A \code{\link[MultiAssayExperiment]{MultiAssayExperiment-class}}
object, or an object inheriting
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}}.}
}
\value{
A named list containing the path(s) to the HDF5 file(s) underlying
each experiment.
}
\description{
getMaeH5paths - extract HDF5 file paths from MAE
}
\author{
Pierre-Luc Germain
}
