% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoadH5mae.R
\name{loadMae}
\alias{loadMae}
\title{loadMae - load a MultiAssayExperiment object}
\usage{
loadMae(filepath, keepAbsPaths = NULL)
}
\arguments{
\item{filepath}{The path to the RDS file containing a
\code{\link[MultiAssayExperiment]{MultiAssayExperiment-class}} object.}

\item{keepAbsPaths}{Logical; whether to keep the absolute paths to HDF5 files
indicated in the object, or to try to update them to the object's location.
By default, the original paths will be preserved if all the HDF5 files are
accessible, and otherwise looked for in the directory of \code{filepath}.}
}
\value{
A \code{\link[MultiAssayExperiment]{MultiAssayExperiment-class}}
object.
}
\description{
loadMae - load a MultiAssayExperiment object
}
