% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainTfModel.R
\name{loadModels}
\alias{loadModels}
\title{Loads models from disk.}
\usage{
loadModels(modelPath)
}
\arguments{
\item{modelPath}{Path of the models .txt-file.}
}
\value{
A list of \link[lightgbm:lightgbm]{lightgbm::lightgbm} models saved on disk .
}
\description{
Loads models saved by \link{saveModels} from disk.
}
