% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextFeatures.R
\name{panContextFeatures}
\alias{panContextFeatures}
\title{Features computed across cellular contexts}
\usage{
panContextFeatures(
  mae,
  nVarSites = 1e+05,
  features = c("MDS_Context", "Max_ATAC_Signal", "ATAC_Variance"),
  annoCol = "context",
  seed = 42,
  saveHdf5 = FALSE,
  outDir = getwd(),
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{mae}{\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} as construced by \code{\link[=prepData]{prepData()}} containing Motif, ATAC-, ChIP-seq,
site-specific features as obtained by \code{\link[=siteFeatures]{siteFeatures()}}.}

\item{nVarSites}{Number of sites with highest ATAC-signal variance to include for the MDS projections.}

\item{features}{Names of features to be added. Can be all or some of "MDS_Context", "Max_ATAC_Signal", "ATAC_Variance".
ChromVAR activity scores and associations to the ATAC-signal will be computed as well if not already present in the experiments of the object.}

\item{annoCol}{Name of column indicating cellular contexts in colData.}

\item{seed}{Integer value for setting the seed for random number generation with \link[base:Random]{base::set.seed}.}

\item{saveHdf5}{If chromVar activity scors and associations experiments should be saved as HDF5 files.}

\item{outDir}{Directory to save HDF5 file to.}

\item{BPPARAM}{Parallel back-end to be used. Passed to \code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}}.}

\item{...}{Arguments passed to \link[chromVAR:getBackgroundPeaks]{chromVAR::getBackgroundPeaks}.}
}
\value{
\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} object with columns added to rowData of the ATAC experiment ("ATAC_Variance", "Max_ATAC_Signal")
or the column data of the ATAC  experiment ("MDS_Context"). If not present in the object experiments for the chromVAR activity scores (name: chromVAR.Activity)
and its association to the ATAC-signal of a site across contexts (name: Activity.Association) are added.
}
\description{
Either site or cellular-context-specific features which are computed based on ATAC-seq data of all contexts combined.
Examples are site-specific variance in the ATAC-signal across contexts or a MDS projection of the cellular contexts.
}
\author{
Emanuel Sonder
}
