% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictTfBinding}
\alias{predictTfBinding}
\title{Predicts transcription-factor binding}
\usage{
predictTfBinding(
  models,
  fm,
  annoCol = "context",
  simplified = TRUE,
  chunk = NULL,
  sparsify = FALSE,
  numThreads = 40,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{models}{List of tree-based gradient boosting \link[lightgbm:lightgbm]{lightgbm::lightgbm} models as obtained by \link{trainTfModel}.}

\item{fm}{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment} object containing features as obtained by \link{getFeatureMatrix}.}

\item{annoCol}{Name of column indicating cellular contexts.}

\item{simplified}{If predictions should be returned as a \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment} object.}

\item{chunk}{Size of chunks if predictions should be performed in chunks to lower the memory footprint. If is \code{NULL} no chunking is applied.}

\item{sparsify}{Should predictions be sparsified. Very small binding probabilities will be rounded to zero.}

\item{numThreads}{Total number of threads to be used. In case \link[BiocParallel:MulticoreParam-class]{BiocParallel::MulticoreParam} or \link[BiocParallel:SnowParam-class]{BiocParallel::SnowParam} with several workers are
are specified as parallel back-ends, \code{floor(numThreads/nWorker)} threads are used per worker.}

\item{BPPARAM}{Parallel back-end to be used. Passed to \code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}}.}
}
\value{
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment} object with predictions of the models saved in the assays (if \code{simplified=TRUE})
and columns corresponding to cellular contexts.
If \code{simplified=FALSE} a \link[Matrix:Matrix]{Matrix::Matrix} is returned with columns corresponding to predictions the models.
}
\description{
Predicts bindings with each model in the bag (as obtained by \link{trainTfModel}) seperately.
}
\author{
Emanuel Sonder
}
