% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictTfBindingBagged}
\alias{predictTfBindingBagged}
\title{Predicts transcription-factor binding}
\usage{
predictTfBindingBagged(
  models,
  featMat,
  chunk = FALSE,
  sparsify = TRUE,
  dataStack = NULL,
  numThreads = 40,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{models}{List of tree-based gradient boosting \link[lightgbm:lightgbm]{lightgbm::lightgbm} models as obtained by \link{trainBagged}}

\item{featMat}{Feature matrix as obtained by \link{getFeatureMatrix} for prediction.}

\item{chunk}{If predictions should be performed on chunks of the data to lower the memory footprint (chunk-size: 1e5)}

\item{sparsify}{Should predictions be sparsified. Very small binding probabilities will be rounded to zero.}

\item{dataStack}{Hold-out data to be used for training stacked models (e.g. if \code{stackingStrat} is either "wMean" or "boostTree").}

\item{numThreads}{Total number of threads to be used. In case \link[BiocParallel:MulticoreParam-class]{BiocParallel::MulticoreParam} or \link[BiocParallel:SnowParam-class]{BiocParallel::SnowParam} with several workers are
are specified as parallel back-ends, \code{floor(numThreads/nWorker)} threads are used per worker.}

\item{BPPARAM}{Parallel back-end to be used. Passed to \code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}}.}
}
\description{
Not done yet
}
