% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictTfBindingStacked}
\alias{predictTfBindingStacked}
\title{Predicts transcription-factor binding}
\usage{
predictTfBindingStacked(modelStacked, featMat, predsBagged = NULL, ...)
}
\arguments{
\item{modelStacked}{Stacked model as obtained by \link{trainStacked}.}

\item{featMat}{Feature matrix to predict on, as obtained by \link{getFeatureMatrix} for prediction.}

\item{predsBagged}{Predictions of the single models in the bag. Only needed if the stacked model has been obtained with the weighted mean strategy.}

\item{...}{Additional arguments passed to \link{predictTfBindingBagged}.}
}
\value{
Matrix with predicted binding probabilities.
}
\description{
Predicts bindings for the stacked model as obtained by \link{trainStacked}
}
