% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepMotifs.R
\name{prepMotifs}
\alias{prepMotifs}
\title{prepMotifs}
\usage{
prepMotifs(
  refCoords,
  motifs,
  outputFolder,
  genome,
  resume = FALSE,
  topPlusCoOnly = TRUE,
  BPPARAM = SerialParam(progress = TRUE)
)
}
\arguments{
\item{refCoords}{Coordinates across which to identify matches. This should be
the same as will be inputted to \code{\link{prepData}}.}

\item{motifs}{A named list of motifs in \code{universalmotif} format, or an object
of class \code{PWMatrixList}. The motif names should be gene names (or whatever
was used to prepare and annotate the ChIP data).}

\item{outputFolder}{A folder where the coordinates of the motif matches will
be stored (one file per motif will be created).}

\item{genome}{The genome to use, either a BSgenome or FaFile object.}

\item{resume}{Logical, whether to only prepare files that aren't already
there.}

\item{topPlusCoOnly}{Logical, whether to report only the top match per
\code{refCoords}, along with the number of other matches. If FALSE, will report
the coordinates of all matches.}

\item{BPPARAM}{BiocParallel BPPARAM object for parallelization.}
}
\value{
A named vector of paths to the files created, which can be used for
input to \code{\link{prepData}}.
}
\description{
Prepare match coordinates files for a list of motifs.
}
\details{
This will save a Rds file for each motif, containing a GRanges of its matches
within the regions defined by \code{refCoords}. If \code{topPlusCoOnly=TRUE}, only the
top match per \code{refCoords} is saved, along with the number of other
significant matches in the regions. If \code{topPlusCoOnly=TRUE}, all matches
passing threshold are saved, and the top per region is flagged using an
\code{isTop} column. Note that to identify the top match per region, a very low
threshold is used in order for most regions to end up with a match.
}
\author{
Pierre-Luc Germain
}
