% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoadH5mae.R
\name{rebaseMaeH5paths}
\alias{rebaseMaeH5paths}
\title{rebaseMaeH5paths - Change the paths of the HDF5 files in a MultiAssayExperiment object.}
\usage{
rebaseMaeH5paths(mae, newBase = vector("character"))
}
\arguments{
\item{mae}{A \code{\link[MultiAssayExperiment]{MultiAssayExperiment-class}}
object.}

\item{newBase}{The new directory in which the HDF5 files are located. This
can either be a single path, or a named vector of paths corresponding to
the different experiments contained in \code{mae}.}
}
\value{
The \code{mae} object with updated HDF5 file links.
}
\description{
rebaseMaeH5paths - Change the paths of the HDF5 files in a MultiAssayExperiment object.
}
