% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoadH5mae.R
\name{saveMae}
\alias{saveMae}
\title{saveMae - Save a MultiAssayExperiment}
\usage{
saveMae(mae, filepath, copyH5 = FALSE)
}
\arguments{
\item{mae}{A \code{\link[MultiAssayExperiment]{MultiAssayExperiment-class}}
object.}

\item{filepath}{The path of the RDS file in which to save the object.}

\item{copyH5}{Logical; whether to make copies of any underlying HDF5 files
in the directory of \code{filepath} and link the saved \code{mae} to those.}
}
\value{
NULL.
}
\description{
saveMae - Save a MultiAssayExperiment
}
