% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainTfModel.R
\name{saveModels}
\alias{saveModels}
\title{Saves models on disk.}
\usage{
saveModels(models, outPath)
}
\arguments{
\item{models}{List of models as obtained by \link{trainTfModel}.}

\item{outPath}{Path for saving the models.}
}
\description{
Saves models obtained by \link{trainTfModel} on disk.
}
