% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfFeatures.R
\name{tfFeatures}
\alias{tfFeatures}
\title{Transcription factor-specific features}
\usage{
tfFeatures(
  mae,
  tfName,
  tfCofactors = NULL,
  features = c("Binding_Patterns", "Promoter_Association", "C_Score",
    "Cooccuring_Motifs", "Cofactor_Binding", "CTCF_Signal", "Associated_Motifs"),
  nPatterns = 50,
  L1 = c(0.5, 0.5),
  nMotifs = 10,
  seed = 42,
  genome = BSgenome.Hsapiens.UCSC.hg38
)
}
\arguments{
\item{mae}{\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} as construced by \code{\link[=prepData]{prepData()}} containing Motif, ATAC-, ChIP-seq
and site-specific features as obtained by \code{\link[=siteFeatures]{siteFeatures()}}.}

\item{tfName}{Name of transcription factor to compute features for.
Provided MultiAssayExperiment object needs to contain ChIP-seq data for the specified transcription factor (check \verb{colData(experiments(mae)$ChIP}).}

\item{tfCofactors}{Names of cofactors (other transcription factors) of the specified transcription factor.}

\item{features}{Names of features to be added. Can be all or some of "Binding_Patterns", "Promoter_Association",
"C_Score", "Cofactor_Binding", "CTCF_Signal", "Cooccuring_Motifs", "Associated_Motifs". Features are stored in the assays of the added experiment.
See \link{listFeatures} for an overview of the features.}

\item{nPatterns}{Number of non-negative matrix factorization (NMF) components to consider for the decomposition of the ChIP-seq peaks matrix.
Passed to \link[RcppML:nmf]{RcppML::nmf}.}

\item{L1}{LASSO penalties for lower rank matrices w and h resulting from the NMF decomposition. Vector with two elements for w and h.
Passed to \link[RcppML:nmf]{RcppML::nmf}}

\item{nMotifs}{Number of associated motifs to select. Chooses \code{nMotifs} co-occuring and \code{nMotifs} mutually exclusive motifs.}

\item{seed}{Integer value for setting the seed for random number generation with \link[base:Random]{base::set.seed}.}

\item{genome}{\link[BSgenome:BSgenome-class]{BSgenome::BSgenome} to be used.}
}
\value{
\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} object with an experiment containing transcription factor-specific features added to \link[MultiAssayExperiment:MultiAssayExperiment-methods]{MultiAssayExperiment::experiments}.
If already an experiment with transcription factor-specific features exists in the object, columns for the specified transcription factor are added to it.
}
\description{
Adds an experiment with features specific for the specified transcription factor, such as co-occuring motifs and binding patterns, to the provided MultiAssayExperiment object.
}
