% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{addATACData}
\alias{addATACData}
\title{Add further ATAC-seq datasets to object}
\usage{
addATACData(
  mae,
  atacData,
  testSet = NULL,
  computeFeatures = FALSE,
  tfName = NULL,
  annoCol = "context",
  shift = FALSE,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{atacData}{Named list of \link[data.table:data.table]{data.table::data.table}/data.frame/\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}
or paths to .bam/.bed files containing ATAC-seq fragments, names being the cellular contexts labels.
Need to contain genomic coordinates (e.g. a chr/seqnames, start and end column).}

\item{testSet}{Vector of cellular context labels used for testing.}

\item{computeFeatures}{If features should be computed (\code{\link[=contextTfFeatures]{contextTfFeatures()}}).}

\item{tfName}{Name of transcription factor to compute context-specific features for.}

\item{annoCol}{Name of column indicating cellular contexts in colData.}

\item{shift}{Only for ATAC-seq data, if Tn5 insertion bias should be considered (only if a strand column is provided).}

\item{BPPARAM}{Parallel back-end to be used. Passed to \code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}}.}

\item{...}{Additional arguments passed to \code{\link[=contextTfFeatures]{contextTfFeatures()}}.}
}
\value{
\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} with added ATAC-seq experiments (and cellular context and TF-specific features if \code{computeFeatures=TRUE}).
}
\description{
Adds additional columns for provided ATAC-seq datasets in the ATAC-seq experiments and
computes features if required.
}
