% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureMatrix.R
\name{getFeatureMatrix}
\alias{getFeatureMatrix}
\title{Feature matrix construction}
\usage{
getFeatureMatrix(
  mae,
  tfName,
  addLabels = TRUE,
  whichCol = c("All", "OnlyTrain", "Col"),
  colSel = NULL,
  norm = c("robust", "min-max", "column", "none"),
  convertInteger = FALSE,
  saveHdf5 = TRUE,
  outDir = NULL,
  prefix = NULL,
  annoCol = "context"
)
}
\arguments{
\item{mae}{\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} as construced by \code{\link[=prepData]{prepData()}} containing Motif, ATAC-, ChIP-seq,
site-specific features as obtained by \code{\link[=siteFeatures]{siteFeatures()}}, transcription factor-specific features as obtained by \code{\link[=tfFeatures]{tfFeatures()}}
and transcription-factor and cellular-context specific features as obtained by \code{\link[=contextTfFeatures]{contextTfFeatures()}}.}

\item{tfName}{Name of transcription factor to compute features for.}

\item{addLabels}{Should ChIP-seq peak labels be added to the feature matrix.}

\item{whichCol}{Should feature matrix be calculated for all cellular contexts (\code{"All"}), only the training data (\code{"OnlyTrain"})
or only for some specific cellular contexts (\code{"Col"}) specified in \code{colSel}.}

\item{colSel}{If \code{whichCol="colSel"}, name of the cellular context to compute the feature matrix for.}

\item{norm}{Normalization strategy to be used for features of different cellular contexts,
can be one of"robust" - substraction of median followed by division by IQR, "min-max" - divison by range between minimum and 0.9-quantile, "column" - division by column sums or "none".
Default is "robust".}

\item{convertInteger}{If feature matrix should be converted to integer (to lower memory footprint).
Not recommend to be used, might lower predictive performance due to loss of information.}

\item{saveHdf5}{If feature matrix should be saved as HDF5 file.}

\item{outDir}{Directory to save HDF5 file to.}

\item{prefix}{Prefix added to filename of feature matrix in case saved as HDF5 file.}

\item{annoCol}{Name of column indicating cellular contexts in colData.}
}
\value{
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment} object with the feature matrix stored in the assays.
Assay is either saved as \link[Matrix:Matrix-class]{Matrix::Matrix} or as \link[HDF5Array:HDF5Array-class]{HDF5Array::HDF5Array} (if \code{saveHdf5=TRUE}).
}
\description{
Compiles feature matrix based on pre-computed features stored in experiments of the provided \link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} object.
}
\details{
Columns in the feature matrix returned are named the following:
\verb{<featureType>_<featureName>} depending of the feature followed by (\verb{<motifSuffix>, <i>, <normedSuffix>}).
}
\author{
Emanuel Sonder
}
