% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInsertionProfiles.R
\name{getInsertionProfiles}
\alias{getInsertionProfiles}
\title{Tn5 insertion counting}
\usage{
getInsertionProfiles(
  atacData,
  motifRanges,
  margin = 200,
  shift = FALSE,
  calcProfile = TRUE,
  profiles = NULL,
  symmetric = FALSE,
  stranded = FALSE,
  subSample = NULL,
  simplified = FALSE,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{atacData}{Named list of \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}, \link[data.table:data.table]{data.table::data.table}, data.frames or paths to .bed /. bam files
containing ATAC-seq fragment coordinates (i.e. chr/seqnames, start, end and optionally a strand column). List names will be used as sample names.
If a single object is provided and it contains a column named "sample", insertion counts will be computed for each sample separately.}

\item{motifRanges}{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object containing coordinates of motif-matches.}

\item{margin}{Margin around motif-matches to consider for computing Tn5 insertion events}

\item{shift}{If Tn5 insertion bias should be considered (only if strand column is provided).}

\item{calcProfile}{If insertion footprint profiles should be computed.}

\item{profiles}{Pre-computed insertion footprint profile to compute weighted insertion counts in case \code{calcProfile=FALSE}.
Needs to contain coordinate (chr/seqnames, start, end) columns and weight column (termed "w").}

\item{symmetric}{If transcription factor footprint profiles should be symmetric around the motif matches. Only used if \code{calcProfile=TRUE}.}

\item{stranded}{If insertion footprint profiles should be computed taking strandedness of fragments into account.}

\item{subSample}{If fragments should be sub-sampled for speed-up.
Default is no sub-sampling, if a number is provided the fragments of each file/object provided will be subsampled to that number.}

\item{simplified}{If return should be simplified and be provided as a \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment} object.}

\item{BPPARAM}{Parallel back-end to be used. Passed to \code{\link[BiocParallel:bpmapply]{BiocParallel::bpmapply()}}.}
}
\value{
If \code{simplified=TRUE} a \link[data.table:data.table]{data.table::data.table} containing insertion counts within and in margins around motif matches and weighted insertion counts in case
an insertion profile is provided or if \code{calcProfile=TRUE}.
If \code{calcProfile=TRUE} also a footprint profile around the motif matches is returned, containing a weight ("w") column corresponding to relative insertion
frequency at the respective position relative to the motif center ("rel_pos").
If \code{simplified=FALSE} a \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment} object is returned with ranges corresponding to the \code{motifRanges} provided as an input.
Assays correspond to the different insertion counts computed, columns will correspond to the samples. If \code{calcProfile=TRUE} the profiles will be saved in the metadata.
}
\description{
Counts Tn5 insertions around provided motif-matches.
If requested also computes insertion footprint profiles and weighted insertion counts.
}
