% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFeatures.R
\name{listFeatures}
\alias{listFeatures}
\title{Overview and description of features.}
\usage{
listFeatures()
}
\value{
\link[data.table:data.table]{data.table::data.table} containing an overview of the features.
}
\description{
Note feature matrix column names and assay names of the MultiAssayExperiment object are the same.
}
