% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Custom MultiAssayExperiment construction.}
\usage{
prepData(
  refCoords,
  motifData,
  atacData,
  chIPData,
  testSet = NULL,
  promoterCoords = NULL,
  aggregationFun = max,
  scoreCol = "score",
  weightCol = NULL,
  isUncertainCol = NULL,
  annoCol = "context",
  shift = FALSE,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{refCoords}{Coordinates across which to aggregate ATAC-, ChIP-seq and motif data.
Constitutes the rowRanges of the RangedSummarizedExperiment object.}

\item{motifData}{Named list of \link[data.table:data.table]{data.table::data.table}/data.frame/\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}
or paths to .bam/.bed files containing motif-matching scores, names being the motif names.
Need to contain genomic coordinates (e.g. a chr/seqnames, start and end column).}

\item{atacData}{Named list of \link[data.table:data.table]{data.table::data.table}/data.frame/\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}
or paths to .bam/.bed files containing ATAC-seq fragments, names being the cellular contexts labels.
Need to contain genomic coordinates (e.g. a chr/seqnames, start and end column).}

\item{chIPData}{Named list of \link[data.table:data.table]{data.table::data.table}/data.frame/\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}
or paths to .bam/.bed files containing ChIP-seq peaks, names being the combination of cellular context and TF labels (e.g. K562_YY1).
Need to contain genomic coordinates (e.g. a chr/seqnames, start and end column) and optionally a column with observational weights (weightCol)
and/or a uncertainty label (isUncertainCol).}

\item{testSet}{Vector of cellular context labels used for testing.}

\item{promoterCoords}{Optionally coordinates of promoters of transcription factor genes.}

\item{aggregationFun}{Function to aggregate}

\item{scoreCol}{Optional, column name of motif-matching data containing matching scores.}

\item{weightCol}{Optional, column name of ChIP-seq data containing observational weights for peaks.}

\item{isUncertainCol}{Optional, column name of ChIP-seq data labelling uncertain peaks (TRUE/FALSE).}

\item{annoCol}{Name of column indicating cellular contexts in colData.}

\item{shift}{Only for ATAC-seq data, if Tn5 insertion bias should be considered (only if a strand column is provided).}

\item{BPPARAM}{Parallel back-end to be used. Passed to \code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}}.}
}
\value{
\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment} with Motif, ATAC- & ChIP-seq experiments.
}
\description{
Creates \link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment::MultiAssayExperiment-class } object for downstream feature construction.
}
